+++ gix/src/lib.rs
@@ -176,7 +176,7 @@ mod attribute_stack;
 pub mod path;
 
 /// The standard type for a store to handle git references.
-pub type RefStore = gix_ref::file::Store;
+pub type RefStore = gix_ref::Store;
 /// A handle for finding objects in an object database, abstracting away caches for thread-local use.
 pub type OdbHandle = gix_odb::memory::Proxy<gix_odb::Handle>;
 /// A handle for finding objects in an object database, abstracting away caches for moving across threads.