+++ gix/src/revision/spec/parse/delegate/revision.rs
@@ -175,7 +175,7 @@ impl delegate::Revision for Delegate<'_> {
     fn nth_checked_out_branch(&mut self, branch_no: usize) -> Result<(), Exn> {
         self.unset_disambiguate_call();
         fn prior_checkouts_iter<'a>(
-            platform: &'a mut gix_ref::file::log::iter::Platform<'static, '_>,
+            platform: &'a mut gix_ref::store::log::iter::Platform<'static, '_>,
         ) -> Result<impl Iterator<Item = (BString, ObjectId)> + 'a, gix_error::Error> {
             match platform.rev().ok().flatten() {
                 Some(log) => Ok(log.filter_map(Result::ok).filter_map(|line| {